Improved code readability. Added threads.py: it contains all the Thread and QThread classes used in pesto.py. Removed pesto_noCmd.py: useless. Created initilize_path function in utilites.py. Created tmp folder: it contains logging files like crashreport.log and temporary files like win_disk_json_example.txt
